remove / fix potentially dangerous panics
