+++ util/strkey/strkey.go
@@ -46,15 +46,6 @@ func Decode(expected VersionByte, src string) ([]byte, error) {
 	return payload, nil
 }
 
-// MustDecode is like Decode, but panics on error
-func MustDecode(expected VersionByte, src string) []byte {
-	d, err := Decode(expected, src)
-	if err != nil {
-		panic(err)
-	}
-	return d
-}
-
 // Encode encodes the provided data to a StrKey, using the provided version
 // byte.
 func Encode(version VersionByte, src []byte) (string, error) {
@@ -78,15 +69,6 @@ func Encode(version VersionByte, src []byte) (string, error) {
 
 	result := base58.FastBase58Encoding(raw.Bytes())
 	return result, nil
-}
-
-// MustEncode is like Encode, but panics on error
-func MustEncode(version VersionByte, src []byte) string {
-	e, err := Encode(version, src)
-	if err != nil {
-		panic(err)
-	}
-	return e
 }
 
 // Version extracts and returns the version byte from the provided source
